Moved colour test to correct folder and package
